pyproject.toml
[tool.poetry]
name = "interbot"
version = "0.1.0"
description = ""
authors = ["ilimk <ilimkanu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.25.1"
python-decouple = "^3.8"
psycopg2 = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
